Update README.md
A group of programs created or adapted for television broadcast with a common series title, usually related to one another in subject or another way.  Often, television series appear once a week during a prescribed time slot; however, they may appear with more or less frequency.  Television series are usually created to be open-ended, not with a predetermined number of episodes.  In a fiction series, the programs typically share the same characters and basic themes.